Managing the array of Recipes using the Service
RecipeService
(1) define and initialize an Array of Services (private access)
(2) define a getRecipes method. (use slice() method to get a copy of the array)
RecipesComponent
(3) Import the file and Add the RecipeService provider to the parent component RecipeComponent
RecipeList
(4) Remove the recipes array initialization because we will use the
(5) In the constructor, Inject the RecipeService that is being inherited from RecipesComponent parent
    constructor(private recipeService: RecipeService)
(6) In the ngOnInit method, Copy the reference of the array of the service in the local array.
    this.recipes = this.recipeService.getRecipes();